+++ robot-description-builder-py/src/link/collision.rs
@@ -1,4 +1,4 @@
-use pyo3::{intern, prelude::*};
+use pyo3::prelude::*;
 use robot_description_builder::{
 	link_data::Collision, linkbuilding::CollisionBuilder, prelude::GroupIDChanger, Transform,
 };
@@ -94,10 +94,7 @@ impl PyCollisionBuilder {
 	}
 
 	pub fn __repr__(&self, py: Python<'_>) -> PyResult<String> {
-		let class_name = py
-			.get_type::<Self>()
-			.getattr(intern!(py, "__qualname__"))?
-			.extract::<&str>()?;
+		let class_name = py.get_type::<Self>().qualname()?;
 
 		let mut data = match self.0.name() {
 			Some(name) => format!("name='{name}', "),
@@ -165,10 +162,7 @@ impl PyCollision {
 	}
 
 	pub fn __repr__(&self, py: Python<'_>) -> PyResult<String> {
-		let class_name = py
-			.get_type::<Self>()
-			.getattr(intern!(py, "__qualname__"))?
-			.extract::<&str>()?;
+		let class_name = py.get_type::<Self>().qualname()?;
 
 		let mut data = match self.inner.name() {
 			Some(name) => format!("name='{name}', "),