+++ robot-description-builder-py/src/link/geometry/cylinder_geometry.rs
@@ -1,4 +1,4 @@
-use pyo3::{basic::CompareOp, intern, prelude::*};
+use pyo3::{basic::CompareOp, prelude::*};
 
 use robot_description_builder::link_data::geometry::{CylinderGeometry, GeometryInterface};
 
@@ -27,10 +27,7 @@ impl PyCylinderGeometry {
 	}
 
 	pub fn __repr__(&self, py: Python<'_>) -> PyResult<String> {
-		let class_name = py
-			.get_type::<Self>()
-			.getattr(intern!(py, "__qualname__"))?
-			.extract::<&str>()?;
+		let class_name = py.get_type::<Self>().qualname()?;
 
 		Ok(format!(
 			"{}({}, {})",