+++ robot-description-builder-py/src/transmission.rs
@@ -119,10 +119,7 @@ impl PyTransmission {
 	}
 
 	pub fn __repr__(&self, py: Python<'_>) -> PyResult<String> {
-		let class_name = py
-			.get_type::<Self>()
-			.getattr(intern!(py, "__qualname__"))?
-			.extract::<&str>()?;
+		let class_name = py.get_type::<Self>().qualname()?;
 
 		let mut data = format!(
 			"'{}', {}, joints=[",