+++ robot-description-builder-py/CHANGELOG.md
@@ -2,13 +2,20 @@
 The Changes between versions of [<b>`robot-description-builder`</b> for <b>Python</b>](https://github.com/SuperJappie08/robot-description-builder/tree/master/robot-description-builder-py#robot-description-builder-) are recored here.
 
 ## Version 0.0.4-dev
+- Migrate to PyO3's new bound API.
+
+### Added
+- Added some tests for the `material` module.
+- Disabled `transmission` module until it is finished.
 
 ### Misc
 - Bumped depency versions:
-    - Rust: 0.20.0
+    - Rust: PyO3 0.20.0 -> 0.21.0
     - Rust: itertools 0.12.0 -> 0.12.1
     - Python: Maturin 1.3.2 -> 1.5.1
     - Python: pytest 7.4.3 -> 8.1.1
+
+
 
 ## Version 0.0.3
 - Convert Collision and VisualBuilder between eachother.